feature/buffer as list and decoding
made the buffer more efficient replacing the buffer with a symmetrical linked list. Also added a braille-reader which uses a decoder for decoding braille symbols to their ascii representation. The decoder uses a binary search tree in which all ascii characters are stored in the leafs. A braille symbol can be stored into a byte in format 0b000001. For every dot a bit is set. This format is used as path in the binary search tree.